requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
flask-cors==4.0.1
python-dotenv==1.0.1
